clean up some file names, get rid of util file by pulling out rank, file and color into their own modules, and peel off zobrist from gamestate.
