+++ chess_engine/src/castle_perm.rs
