+++ chess_engine/src/lib.rs
@@ -2,10 +2,13 @@
 #![allow(clippy::module_inception)]
 #![allow(unused)]
 pub mod board;
-pub mod castle_perms;
+pub mod castle_perm;
+pub mod color;
 pub mod error;
+pub mod file;
 pub mod gamestate;
 pub mod moves;
-pub mod pieces;
-pub mod squares;
-pub mod util;
+pub mod piece;
+pub mod rank;
+pub mod square;
+pub mod zobrist;
